Load all the objects by stepping the webservice repeatedly.
Yes, this will be slow for large lists! We might consider inserting a pager in the frontend.
But it is nicer than the current loan lists missing some titles when the limit of 50 uncached objects are hit.
opensearch limits stepValue to 50 (even though it doesn't say so in the documentation)...